+++ R/DuckDBAtomicList-class.R
@@ -281,7 +281,9 @@ setMethod("as.list", "DuckDBAtomicList", function(x, use.names = TRUE) {
     # Columns are ordered: datacol (column 1), then keycol (column 2)
     names <- .map_keycol_names(x@table@keycols[[1L]], df[[2L]])
     result <- setNames(df[[1L]], names)
-    result <- result[rownames(x@table)]
+    if (!.has_row_number(x@table)) {
+        result <- .reindexByStoredKeys(result, rownames(x@table))
+    }
 
     if (!use.names) {
         names(result) <- NULL